+++ chess.py
@@ -19,7 +19,9 @@ class Chess:
             raise EmptyPosition()
         if not piece.get_color() == self.__turn__:
             raise InvalidTurn()
-        possible=piece.valid_positions(from_row,from_col)
+        if type(piece).__name__=='King':
+            possible=piece.valid_positions(from_row,from_col,for_all_moves=False)
+        else: possible=piece.valid_positions(from_row,from_col)
         if not piece.is_row_col_in_valid_positions(to_row,to_col,possible):
             raise InvalidMove()
         self.__board__.move(from_row, from_col, to_row, to_col)