+++ tests/test_king.py
@@ -2,6 +2,7 @@ import unittest
 from king import King
 from board import Board
 from pawn import Pawn
+from knight import Knight
 
 class TestKing(unittest.TestCase):
 
@@ -26,7 +27,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_va(4,4)
         self.assertEqual(possibles,[(3,4)])
 
-    def test_move_vertical_ascendant_with_enemy_piece(self):
+    def test_move_white_piece_vertical_ascendant_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -34,11 +35,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_va(4,4)
         self.assertEqual(possibles,[(3,4)])
 
-    def test_move_vertical_ascendant_with_own_piece(self):
+    def test_move_black_piece_vertical_ascendant_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(3,4,Pawn("WHITE", board))
+        possibles=king.possible_positions_va(4,4)
+        self.assertEqual(possibles,[(3,4)])
+
+    def test_move_white_piece_vertical_ascendant_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(3,4,Pawn("WHITE", board))
+        possibles=king.possible_positions_va(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_vertical_ascendant_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(3,4,Pawn("BLACK", board))
         possibles=king.possible_positions_va(4,4)
         self.assertEqual(possibles,[])
 
@@ -51,7 +68,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_vd(4,4)
         self.assertEqual(possibles,[(5,4)])
 
-    def test_move_vertical_descendant_with_enemy_piece(self):
+    def test_move_white_piece_vertical_descendant_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -59,11 +76,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_vd(4,4)
         self.assertEqual(possibles,[(5,4)])
 
-    def test_move_vertical_descendant_with_own_piece(self):
+    def test_move_black_piece_vertical_descendant_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(5,4,Pawn("WHITE", board))
+        possibles=king.possible_positions_vd(4,4)
+        self.assertEqual(possibles,[(5,4)])
+
+    def test_move_white_piece_vertical_descendant_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(5,4,Pawn("WHITE", board))
+        possibles=king.possible_positions_vd(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_vertical_descendant_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(5,4,Pawn("BLACK", board))
         possibles=king.possible_positions_vd(4,4)
         self.assertEqual(possibles,[])
 
@@ -76,7 +109,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_hr(4,4)
         self.assertEqual(possibles,[(4,5)])
 
-    def test_move_horizontal_right_with_enemy_piece(self):
+    def test_move_white_piece_horizontal_right_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -84,11 +117,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_hr(4,4)
         self.assertEqual(possibles,[(4,5)])
 
-    def test_move_horizontal_right_with_own_piece(self):
+    def test_move_black_piece_horizontal_right_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(4,5,Pawn("WHITE", board))
+        possibles=king.possible_positions_hr(4,4)
+        self.assertEqual(possibles,[(4,5)])
+
+    def test_move_white_piece_horizontal_right_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(4,5,Pawn("WHITE", board))
+        possibles=king.possible_positions_hr(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_horizontal_right_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(4,5,Pawn("BLACK", board))
         possibles=king.possible_positions_hr(4,4)
         self.assertEqual(possibles,[])
 
@@ -101,7 +150,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_hl(4,4)
         self.assertEqual(possibles,[(4,3)])
 
-    def test_move_horizontal_left_with_enemy_piece(self):
+    def test_move_white_piece_horizontal_left_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -109,11 +158,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_hl(4,4)
         self.assertEqual(possibles,[(4,3)])
 
-    def test_move_horizontal_left_with_own_piece(self):
+    def test_move_black_piece_horizontal_left_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(4,3,Pawn("WHITE", board))
+        possibles=king.possible_positions_hl(4,4)
+        self.assertEqual(possibles,[(4,3)])
+
+    def test_move_white_piece_horizontal_left_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(4,3,Pawn("WHITE", board))
+        possibles=king.possible_positions_hl(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_horizontal_left_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(4,3,Pawn("BLACK", board))
         possibles=king.possible_positions_hl(4,4)
         self.assertEqual(possibles,[])
 
@@ -126,7 +191,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_dar(4,4)
         self.assertEqual(possibles,[(3,5)])
 
-    def test_move_diagonal_ascendant_right_with_enemy_piece(self):
+    def test_move_white_piece_diagonal_ascendant_right_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -134,11 +199,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_dar(4,4)
         self.assertEqual(possibles,[(3,5)])
 
-    def test_move_diagonal_ascendant_right_with_own_piece(self):
+    def test_move_black_piece_diagonal_ascendant_right_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(3,5,Pawn("WHITE", board))
+        possibles=king.possible_positions_dar(4,4)
+        self.assertEqual(possibles,[(3,5)])
+
+    def test_move_white_piece_diagonal_ascendant_right_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(3,5,Pawn("WHITE", board))
+        possibles=king.possible_positions_dar(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_diagonal_ascendant_right_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(3,5,Pawn("BLACK", board))
         possibles=king.possible_positions_dar(4,4)
         self.assertEqual(possibles,[])
 
@@ -151,7 +232,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_dal(4,4)
         self.assertEqual(possibles,[(3,3)])
 
-    def test_move_diagonal_ascendant_left_with_enemy_piece(self):
+    def test_move_white_piece_diagonal_ascendant_left_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -159,11 +240,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_dal(4,4)
         self.assertEqual(possibles,[(3,3)])
 
-    def test_move_diagonal_ascendant_left_with_own_piece(self):
+    def test_move_black_piece_diagonal_ascendant_left_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(3,3,Pawn("WHITE", board))
+        possibles=king.possible_positions_dal(4,4)
+        self.assertEqual(possibles,[(3,3)])
+
+    def test_move_white_piece_diagonal_ascendant_left_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(3,3,Pawn("WHITE", board))
+        possibles=king.possible_positions_dal(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_diagonal_ascendant_left_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(3,3,Pawn("BLACK", board))
         possibles=king.possible_positions_dal(4,4)
         self.assertEqual(possibles,[])
 
@@ -176,7 +273,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_ddr(4,4)
         self.assertEqual(possibles,[(5,5)])
 
-    def test_move_diagonal_descendant_right_with_enemy_piece(self):
+    def test_move_white_piece_diagonal_descendant_right_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -184,11 +281,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_ddr(4,4)
         self.assertEqual(possibles,[(5,5)])
 
-    def test_move_diagonal_descendant_right_with_own_piece(self):
+    def test_move_black_piece_diagonal_descendant_right_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(5,5,Pawn("WHITE", board))
+        possibles=king.possible_positions_ddr(4,4)
+        self.assertEqual(possibles,[(5,5)])
+
+    def test_move_white_piece_diagonal_descendant_right_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(5,5,Pawn("WHITE", board))
+        possibles=king.possible_positions_ddr(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_diagonal_descendant_right_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(5,5,Pawn("BLACK", board))
         possibles=king.possible_positions_ddr(4,4)
         self.assertEqual(possibles,[])
 
@@ -201,7 +314,7 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_ddl(4,4)
         self.assertEqual(possibles,[(5,3)])
 
-    def test_move_diagonal_descendant_left_with_enemy_piece(self):
+    def test_move_white_piece_diagonal_descendant_left_with_enemy_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
@@ -209,11 +322,27 @@ class TestKing(unittest.TestCase):
         possibles=king.possible_positions_ddl(4,4)
         self.assertEqual(possibles,[(5,3)])
 
-    def test_move_diagonal_descendant_left_with_own_piece(self):
+    def test_move_black_piece_diagonal_descendant_left_with_enemy_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(5,3,Pawn("WHITE", board))
+        possibles=king.possible_positions_ddl(4,4)
+        self.assertEqual(possibles,[(5,3)])
+
+    def test_move_white_piece_diagonal_descendant_left_with_own_piece(self):
         board=Board(for_test=True)
         king=King('WHITE',board)
         board.set_piece(4,4,king)
         board.set_piece(5,3,Pawn("WHITE", board))
+        possibles=king.possible_positions_ddl(4,4)
+        self.assertEqual(possibles,[])
+
+    def test_move_black_piece_diagonal_descendant_left_with_own_piece(self):
+        board=Board(for_test=True)
+        king=King('BLACK',board)
+        board.set_piece(4,4,king)
+        board.set_piece(5,3,Pawn("BLACK", board))
         possibles=king.possible_positions_ddl(4,4)
         self.assertEqual(possibles,[])
 
@@ -230,6 +359,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(1,5,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(5,4),(4,5),(4,3),(3,5),(3,3),(5,5),(5,3)])
 
@@ -237,6 +367,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(7,3,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(4,5),(4,3),(3,5),(3,3),(5,5),(5,3)])
 
@@ -244,6 +375,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(5,7,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(4,3),(3,5),(3,3),(5,5),(5,3)])
 
@@ -251,6 +383,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(5,1,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(4,5),(3,5),(3,3),(5,5),(5,3)])
 
@@ -258,6 +391,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(1,6,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(4,5),(4,3),(3,3),(5,5),(5,3)])
 
@@ -265,6 +399,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(1,2,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(4,5),(4,3),(3,5),(5,5),(5,3)])
 
@@ -272,6 +407,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(7,6,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(4,5),(4,3),(3,5),(3,3),(5,3)])
 
@@ -279,6 +415,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(7,2,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(4,5),(4,3),(3,5),(3,3),(5,5)])
 
@@ -286,6 +423,8 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(1,5,Knight("WHITE", board))
+        board.set_piece(7,3,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(4,5),(4,3),(3,5),(3,3),(5,5),(5,3)])
 
@@ -293,5 +432,7 @@ class TestKing(unittest.TestCase):
         board=Board(for_test=True)
         king=King('BLACK',board)
         board.set_piece(4,4,king)
+        board.set_piece(5,7,Knight("WHITE", board))
+        board.set_piece(5,1,Knight("WHITE", board))
         possible=king.valid_positions(4,4)
         self.assertEqual(possible,[(3,4),(5,4),(3,5),(3,3),(5,5),(5,3)])